q1_rag_vs_sqlAgent/requirements.txt
# Core dependencies
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.24.3
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1

# RAG System dependencies
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.5
openai==1.6.1
chromadb==0.4.22
sentence-transformers==2.2.2
faiss-cpu==1.7.4

# SQL Agent dependencies
langchain-experimental==0.0.49
sqlalchemy-utils==0.41.1

# Evaluation and benchmarking
timeit==1.0.0
memory-profiler==0.61.0
psutil==5.9.6

# Web framework for demo
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Utilities
tqdm==4.66.1
rich==13.7.0
click==8.1.7
faker==20.1.0 